[tests] align names with tested modules
AI assistance was used for this change.
