+++ tests/utils/reward_score/test_unified_reward_on_cpu.py
@@ -1,0 +1,23 @@
+# Copyright 2026 Bytedance Ltd. and/or its affiliates
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""CPU tests for the UnifiedReward parser."""
+
+from verl_omni.utils.reward_score.unified_reward import _parse_unified_reward_scores
+
+
+def test_unified_reward_requires_all_labeled_axes():
+    assert _parse_unified_reward_scores("Alignment Score: 4\nCoherence Score: 5") == {}
+    assert _parse_unified_reward_scores(
+        "Alignment Score: 4\nCoherence Score: 5\nStyle Score: 3"
+    ) == {"alignment": 4.0, "coherence": 5.0, "style": 3.0}
